Working Scanner in Scanner.java. Demo in Demo.java
