--- output1.txt
@@ -1,6 +1,0 @@
-Lexeme: print	Token: reserved word
-Lexeme: (	Token: bracket
-Lexeme: "	Token: bracket
-Lexeme: Hello, World!	Token: user variable
-Lexeme: "	Token: bracket
-Lexeme: )	Token: bracket
